requirements.txt
streamlit==1.38.0
spotipy==2.24.0
python-dotenv==1.0.1